console: add scrolling (#13)
depends on #12

this branch adds support for scrolling the task list. the current task
at the scroll position is highlighted, which also gives us a way to
select tasks for future features.

here's a quick demo. since the scrolling is quite simple, the demo also
shows off some nice error handling (i kill the server we're connected
to).

[![asciicast](https://asciinema.org/a/412116.svg)][1]

[1]: https://asciinema.org/a/412116