subscriber: fix busy loop in aggregator task
currently, there's an issue where the `console-subscriber` background
aggregator task busy-loops, using 100% CPU time. this is because it's
currently awaiting a `recv` call on the events channel on every
iteration of the aggregate loop...and events are generated when a task
is exited. this includes the aggregator task! so, when the aggregator
task has finished a poll, the subscriber sends an exit message...which
wakes the aggregator task back up.

this branch fixes the issue by changing the aggregator task to *not*
include new messages on the events channel in the `select!` that will
wake up the aggregator. instead, it allows messages to be buffered, and
drains all events in the channel whenever it's woken by something else.
to avoid losing events due to a full buffer, i've also added a "high
water line" where the `TasksLayer` will notify the aggregator task after
a send, *if* the channel is approaching its maximum capacity (with some
wiggle room). now, the aggregator is woken in the following cases:

* the data flush interval has elapsed
* a new client has subscribed
* the event buffer is nearing capacity

and the buffer is drained before going back to sleep in all these cases.
this should avoid the busy loop and prevent the aggregator from
monopolizing CPU time.

we can probably also shrink the event buffer a bit now that we're always
woken when nearing capacity...