feat(console): add Task Details view (#24)
This starts adding the concept of there being multiple possible "views"
of the console. To add a second view, this includes a basic "task
details" view. If you hit "enter" when scrolling through the task list,
it will go to the details view. It's pretty bare-bones for now, just
showing the same info you would see in the task list view, but with a
different layout. The point of opening this now is to simply start the
process of having multiple views.

Some things we _could_ show are listed in [this doc][1]. I'm thinking
about another PR soon that would include a list of logs (tracing events)
scoped to the task to show on the details page, so you can see the logs
of just a single task easily.

[1]: https://hackmd.io/a4P75rj0RP6qettxAVzv8w?view#Task-instance
- Bold the labels
- Make Esc go back to the task list view
- Fix off-by-one selected_task when sort ascending
- Add task metadata fields to output

Co-authored-by: Eliza Weisman <eliza@buoyant.io>