fix: detect completed tasks even if console connects after (#29)
The "completion" of a task would result in that being sent to all
watchers in a task update. However, after a flush, that status wasn't
stored in a way that new watchers could detect. This changes it so that
a task is determined "completed" when it includes a "total" duration,
instead of being in a separate part of the `TasksUpdate` proto message.