feat(subscriber): drop data for completed tasks  (#31)
Currently, the console subscriber will store task data for tasks even
after they complete. This means that it will use an increasing amount of
memory over time. This branch updates the aggregator task to drop data
for completed spans after a configurable retention period.

I've also added a builder for configuring the console subscriber, including
changing the retention period. By default, completed tasks are retained for
an hour, so that some historical data can be played back, but this can be
reduced to decrease memory usage.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>