feat(console): nicer poll time layout in details view
This branch slightly improves the layout of the poll time stats in the
task details view. Now, instead of doing a 50/50 horizontal split that
requires us to break the percentiles into two columns, we do an
asymmetrical horizontal split, and make the poll time area two lines
longer. This allows us to fit the percentiles in a single column, and
lets us draw a slightly bigger histogram (hopefully making it a little
nicer to read).

This also simplifies the code a little.