feat(console): nicer poll time layout in details view (#58)
This branch slightly improves the layout of the poll time stats in the
task details view. Now, instead of doing a 50/50 horizontal split that
requires us to break the percentiles into two columns, we do an
asymmetrical horizontal split, and make the poll time area two lines
longer. This allows us to fit the percentiles in a single column, and
lets us draw a slightly bigger histogram (hopefully making it a little
nicer to read).

This also simplifies the code a little, since we can remove the code
for splitting the poll time histogram area into two columns.

Before:
![image](https://user-images.githubusercontent.com/2796466/124795284-b71e1f80-df04-11eb-87f4-6bb4e5b93518.png)

After:
![image](https://user-images.githubusercontent.com/2796466/124795123-96ee6080-df04-11eb-839c-c1cd5be8a329.png)