feat(subscriber): support multiple task callsites (#68)
This branch updates the console subscriber to allow recording multiple
`tracing` span callsites as tasks, rather than only tracking the first
callsite that's found. This will allow collecting data from multiple
runtimes, and also removes the requirement that all task spans originate
from a single callsite, which will make implementing runtime
instrumentation simpler.

This is implemented by adding a new data structure for storing a set of
callsites. It's implemented as an array of `AtomicPtr`s to metadata,
plus an `AtomicUsize` tracking the next index to push to. As callsites
are registered, their addresses are stored at the push index, which is
incremented. When testing if a callsite is in the set, the subscriber
performs a linear scan over the pushed callsites.

A linear scan over an array is not *algorithmically* optimal, but in
practice, it is probably the best data structure here. Although the
algorithmic complexity of doing a linear scan is O(_n_), and using a
`HashSet` or something is O(1), there are never going to be a great deal
of items in the array --- in many programs, each callsite list will only
ever have a single entry, and in extreme cases, like when using `tokio`,
`rayon`, and `async-std` in the same program, there might be four or
five entries. So, the cost of an O(_n_) linear scan over a very small
_n_ slice is likely much lower than the constant-factor costs of using
an O(1) `HashSet` in this case. Also, the array allows us to avoid using
a lock, or introducing the additional overhead of a concurrent hash set.

The same data structure is used to track waker op callsites, so we can
now avoid doing string comparisons to detect those callsites.

In addition, I've updated the console to display the metadata targets of
tasks. This is useful because tasks may now be recorded from callsites
in different libraries, or in different modules of one library (e.g.
tokio's task, blocking task, and local task spans could now have
separate callsites). The targets can be used to distinguish between
different task span locations.

Here's a screenshot showing a demo app that uses `tokio`'s normal _and_
blocking tasks, alongside a (work in progress) version of `rayon` with
console instrumentation:
![image](https://user-images.githubusercontent.com/2796466/125658047-aabf85ea-aa1f-485a-8e55-5c8ef67168fd.png)

Signed-off-by: Eliza Weisman <eliza@buoyant.io>