feat(console): color durations based on units  (#72)
This branch adds color-coding for durations based on their unit, as
described in #71. The colors are a simple gradient, from picoseconds to
whole seconds.

Currently, the unit is detected by formatting the value using the
`fmt::Debug` impl for `Duration`, and looking at the last two characters
to determine the unit. Admittedly, this is kind of janky, but it seemed
like re-implementing all of the `Duration` formatting logic would be
worse.

Right now, this assumes the terminal supports the ANSI 256-color
palette. I'm working on an additional change to detect what color
palettes the terminal supports and select colors based on that, but I
figured I'd do that in a separate PR.

Here's a little example screenshot:

![image](https://user-images.githubusercontent.com/2796466/126686509-e5fe37f7-82e7-422e-a486-c8d60e97ed6a.png)

Signed-off-by: Eliza Weisman <eliza@buoyant.io>