feat(console): don't display fields with "" values (#83)
Depends on #81 (since it touches the same code and would cause merge conflicts).

* feat(console): don't display fields with "" values

Now that Tokio supports optional user-provided task names
(in tokio-rs/tokio#3881), the task spans it generates may contain empty
fields (if the user does not provide a name for a spawned task).
However, the console always displays the `task.name` field, even when
they are empty. This wastes space, and looks kind of ugly.

This branch updates the console to skip displaying fields whose value is
the empty string.

### Before:
![image](https://user-images.githubusercontent.com/2796466/128227118-7ee3b4d7-634f-410f-a1d0-4e5afd423566.png)

### After:
![image](https://user-images.githubusercontent.com/2796466/128227203-248bd059-839a-4254-8470-2761f726c885.png)

Signed-off-by: Eliza Weisman <eliza@buoyant.io>

* refac(console): factor out field from pb conversion

This factors out the protobuf `Field` -> internal `Field` conversion
into a method, and cleans it up a bit. No functional change, except that
we now take the proto `Field` type by value, so we don't have to clone
strings out of it like we did previously (due to taking it by
reference). I also added the metadata ID to the `Metadata` type, so that
we can continue asserting that the metadata's ID matches the expected
one for the field, without having to pass that in separately.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>