feat(console): replace tasks kind column with state column
This branch replaces the "kind" column in the tasks list view with a new
"state" column that indicates whether the task is currently running,
idle, or terminated. The "kind" column was initially intended to
distinguish between Tokio's blocking tasks and async tasks, but the
task's fields column also conveys this information, so the "kind" column
was really just useful for indicating whether the task was active or
terminated. Distinguishing vs sleeping and idle tasks here makes this
column somewhat more useful.

This currently uses the Unicode "play", "pause", and "stop" symbols, as
proposed in #25, to display the different task states. In a subsequent
change, I'll add support for detecting and determining whether UTF-8 is
supported by the terminal.