feat(console): comprehensive color configuration (#88)
This branch updates the console application to:
- Detect what colors are supported by the current terminal and choose
  them accordingly
- Allow user settings to override the color palette
- Allow user settings to override whether particular UI elements are
  colored

This change is rather complicated. It introduces `clap` for CLI
arguments, to allow the user to override the color palette.

When the color palette is _not_ overridden, the console will determine
what colors the terminal supports via the following process:
- If the `COLORTERM` environment variable is set to "24bit" or
  "truecolor", this indicates that the terminal supports 24-bit RGB
  color. In that case, all colors are used.
- Otherwise, the console attempts to use `tput colors` to query the
  `terminfo` database to see what colors the current terminal supports.
- If the call to `tput colors` fails, we assume no colors are supported.

Additionally, new CLI args allow the user to toggle whether durations
are color-coded, and whether terminated tasks are dimmed. In the future,
we will probably also want to add support for parsing these from a
config file --- in general, users probably won't want to set these
preferences from the command line every time they run the console.
However, I decided to punt on adding a config file in this PR --- the
addition of new CLI arg parsing logic is a big enough change. We can add
a simple TOML file or something in a follow-up change.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>