feat(console): replace tasks kind column with state column  (#89)
This branch replaces the "kind" column in the tasks list view with a new
"state" column that indicates whether the task is currently running,
idle, or terminated. The "kind" column was initially intended to
distinguish between Tokio's blocking tasks and async tasks, but the
task's fields column also conveys this information, so the "kind" column
was really just useful for indicating whether the task was active or
terminated. Distinguishing vs sleeping and idle tasks here makes this
column somewhat more useful.

The task list can also be sorted by the state column. We may want to
consider making this the default sorting in the future...

This currently uses the Unicode "play", "pause", and "stop" symbols, as
proposed in #25, to display the different task states. In a subsequent
change, I'll add support for detecting and determining whether UTF-8 is
supported by the terminal.

![image](https://user-images.githubusercontent.com/2796466/129458704-a5643d93-6fe1-492f-aafe-6f70a8338b78.png)

Signed-off-by: Eliza Weisman <eliza@buoyant.io>