chore(docs): fix crashy asciicast in readme
The asciinema SVG in the readme triggers a bug in recent versions of
Firefox that makes performance extremely bad while viewing the readme.
This is quite annoying because it makes the browser really slow every
time you're looking at the repo's main page in firefox.

Also, the asciinema shows a pretty outdated version of the console UI.

This commit replaces the asciinema SVG embed with some more recent
screenshots.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>