feat(subscriber): periodically shrink task data
This adds periodic `HashMap::shrink_to_fit` calls to the console
subscriber's `TaskData` hashmaps. This should prevent unbounded memory
growth in long-running applications.

To reduce the performance impact of periodically reallocating, we only
shrink the hash maps occasionally. The interval is probably tuneable,
but for now, I just picked something arbitrary.

In the future we could probably try to implement more advanced
heuristics for shrinking. For example, we could track an average size,
and only shrink when we go significantly over that. But, for now, this
is probably fine.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>