feat(subscriber): accept durations with units
This branch updates the `console-subscriber` env var parsing code to accept
durations with units. Durations may now be given in any of `ns`, `us`,
`ms`, `s`, `sec`, `seconds`, `m`, `min`, `minutes`, `h`, or `hours`.
Durations specified in seconds or a larger unit may also be given as fractional values,
such as `1.5min`.

As a follow-up, we may also want to allow specifying durations as a
combination of multiple units, such as `3h15m`.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>