feat(console): periodically shrink growable collections (#94)
The `Aggregator` task spawned by the console subscriber contains a
number of growable collections (`HashMap`s and `Vec`s) that are used to
store aggregated task data, as well as handles for RPC channels. These
will grow over time as more values are added. In a long-running program,
they may eventually become quite large. Because the Rust standard
library's collections have amortized reallocation behavior, the memory
use of a `HashMap` or `Vec` is _not_ O(_n_), where _n_ is the current
number of elements. Instead, the memory use is O(max(_n_)) --- the
_largest_ number of elements that have _ever_ been in that collection
over the entire lifetime of the program.

In some cases, such as when a bursty workload results in spikes in the
number of spawned tasks, the allocated capacity of these collections may
be significantly higher than the actual amount of space needed to store
the current values in the collection.

In order to avoid unbounded memory growth when using the console
subscriber, this branch adds periodic `shrink_to_fit` calls to the
`Aggregator` task's `Vec`s and `HashMap`s. Of course, we don't want to
call `shrink_to_fit` every time elements are removed from a collection
--- this would negate the performance benefits of amortized
reallocation, since we would constantly be reallocating and coping the
collection into new allocations. Instead, we want to have some kind of
heuristic that ensures we shrink collections occasionally to avoid
unbounded growth, but don't do it so often that the program spends all
its time `malloc`ing and copying. 

This branch currently implements a fairly simple heuristic where we
decide to shrink based on two factors: a counter and interval between
shrinks, and a minimum number of bytes necessary to shrink. The way this
works is: every time we call `retain` on a collection and actually
remove elements, we increment the counter. If the count is greater than
the minimum interval between shrinks, *and* the difference between the
collection's allocated capacity and its actual used size (the amount of
space that would be freed by shrinking) is over the minimum number of
bytes, we call `shrink_to_fit` and reset the counter. Otherwise, we keep
going. The counter and interval ensure that we only shrink occasionally,
instead of after every `retain` call, while the minimum number of bytes
ensures we don't call `shrink_to_fit` when we are close to the current
allocated capacity and would only free up a small amount of storage
anyway.

The default values for these constants are currently an interval of at
least 60 retains between shrinks, and a minimum freed bytes of 4kb. Both
of these values were chosen totally arbitrarily and there's definitely a
lot of room for tuning them, but I just wanted to pick something for
now.

In the future, we may want to consider a smarter heuristic, such as
tracking a moving average of collection capacity and length, and
shrinking back to the average capacity after we have returned to the
average _length_ after a significant deviation in length. The idea
behind this that we  allow the collection to "warm up" to the maximum
size needed for "normal" program operation, and stay there without
causing unnecessary reallocations when everything is normal. If,
however, we have to grow the collection significantly to compensate for
a sudden spike in the number of tasks, we would eventually return to the
size where normal operation doesn't require us to grow the collection.
However, this kind of approach is significantly more complex than the
simple heuristic in this PR, so I thought we ought to save that for
later.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>