subscriber: remove trace event calls from the subscriber (#95)
It was noted that emitting a tracing event inside the subscriber could
trigger a deadlock, since `std::sync::Once` is not re-entrant. This
removes the tracing from the subscriber `Layer` methods.

If the capacity was full, sending the event was spawned into tokio, but
the event might not have been inside a runtime in the first place. And
it essentially turns it into an unbounded queue. So that is also
removed.

It also changes from using a constant `FLUSH_AT_CAPACITY` to a struct
field that is based on the configured event buffer capacity. A value of
half the buffer capacity was chosen, since hitting a full channel will
now mean the event is completely lost.

Closes #27