feat: add pause and resume (#78)
A first step to address #70. You can now press the spacebar to "pause"
the console. The subscriber will continue to process events, but as
along as the client is "paused", it won't send any task updates.
Resuming will bring the client back to "live", getting all updates
again.

![paused](https://user-images.githubusercontent.com/51479/127583646-0cb879c9-0fb7-4025-90c8-4edee8c3e08d.PNG)

I've tried a few different UI things, from making the whole background
of the top red, to just making the word red... It's fine if others feel
there's a better or prettier way to show this, but to keep moving
forward, I just picked something that seemed good enough.

Closes #85