fix(console): don't crash if `COLORTERM`/`LANG` is not set (#98)
Currently, the console CLI/env var handling code _requires_ the
`COLORTERM` env variable. This isn't correct, as some terminals may not
set it. In that case, we should assume the terminal isn't advertising
24-bit color support, rather than crashing.

Currently, this happens if I unset the env variable:

```console
$ unset COLORTERM

$ cargo run
    Finished dev [unoptimized + debuginfo] target(s) in 0.32s
     Running `target/debug/console`
error: The following required arguments were not provided:
    --colorterm <truecolor>

USAGE:
    console [FLAGS] [OPTIONS] --lang <LANG> --colorterm <truecolor> [TARGET_ADDR]

```

After this branch, if I unset `COLORTERM`, the console just runs without
colors.

Similarly, the `LANG` environment variable is no longer required, either.
If it's unset, we'll default to `en_us.UTF-8`. Possibly, we should *really*
be falling back to ASCII if `$LANG` is not set, but _most_ terminals really
ought to support UTF-8...

Finally, this fixes a bug in the way we parse `tput colors` output. The output
ends in a newline, but our parsing code doesn't handle whitespace...so we 
will currently always assume no colors are supported if we fall back to `tput`.
This branch adds a `trim` call.