console: prevent huge take names from panicking
If a task name were too long, calculating the remaining width for the
fields column would panic on a subtraction overflow. This removes the
need to calculate the remainder, by just setting a percentage constraint
instead.

This also caps the `view::Width` so that an excessively long string
doesn't overflow the stored `u16`.