fix(console): don't panic when scrolling an empty task list (#101)
Currently, if there are no tasks in the task list, the console will
panic when using the arrow keys to scroll the selected index in the
table. This is because the code for scrolling the table always subtracts
1 from the length of the list of tasks when determining the last index.
This overflows, resulting in a panic (in debug mode).

This branch fixes that by only attempting to scroll the selected task
index if the length of the task list is greater than 0. I factored out
some duplicate logic for handling this case out into a separate method,
and changed `scroll_next` and `scroll_prev` to call this method. We
still perform the same potentially overflowing subtractions, but they
are now guarded by an `if` that returns early if the length is 0.

I considered just using saturating subtraction to fix the panic, but if
we just changed the subs to saturating, we would still try to "scroll"
the task list, indenting it slightly to show the selected task cursor.
This seems unnecessary when we have no tasks to select, so I thought it
was nicer to just not set the selected index at all in this case.

I did some testing by modifying the example app to not spawn any tasks.
The console no longer crashes after this change.

Fixes #100

Co-authored-by: Sean McArthur <sean@seanmonstar.com>