feat: emit and show self-wake counts for tasks (#112)
This counts all wakes of a task that happen while the task is "currently
polled". This could be from `task::yield_now()`, from `tokio::coop`, or
any other pattern that triggers the task to wake even though it's
currently being polled.

This implementation builds upon @seanmonstar's branch #55. However,
rather than detecting self-wakes by simply looking at _whether_ the task
is being polled at all, we instead detect them by checking whether the
task's span is on the current thread's stack of entered spans. This
approach should _not_ produce self-positives when a task is woken by
another thread while it is currently entered (as discussed in
https://github.com/tokio-rs/console/pull/55#discussion_r659223228).

Testing this with the "burn" subcommand for the example app, which
spawns one task that does a lot of self-wakes via `task::yield`,
indicates that it is, in fact, working properly. Here are the stats for the
`burn` task:
![image](https://user-images.githubusercontent.com/2796466/132136922-2d0d9fbd-e8eb-475c-8ab8-855f8fce5127.png)
Note the very large number of self-wakes.

In comparison, here are the task stats for a "normal" task, which does
not wake itself (and is instead woken by timers):
![image](https://user-images.githubusercontent.com/2796466/132136964-514d7560-074a-41b1-9d6b-621c0dfac935.png)
Note that no self-wakes stat is displayed.

Closes #55.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>
Co-authored-by: Sean McArthur <sean@seanmonstar.com>