feat(console): add name to task details
Currently, the task details view doesn't actually show the task's name.
This commit changes it to add the name if the task has one.

I also changed `Task::name` to return an `Option` so that we can more
easily detect whether or not the task has a name (rather than checking
if the string is empty).