feat(subscriber): name tasks spawned by the console subscriber
This branch updates the `console-subscriber` crate so that the
"aggregate" and "serve" tasks spawned by the console are spawned with
names. This allows the user to distinguish between tasks spawned by the
console subscriber and tasks spawned by other parts of the application.

Note that this does *not* include the tasks spawned by Tonic to handle
each accepted client connection. Those tasks are spawned inside of
Tonic, rather than by the console-subscriber crate, so we can't easily
add our own names to them. We could probably fix that by using Tonic's
lower level APIs to spawn our own per-connection tasks, and give them
names...but that seems like a follow-up PR.

This is a first pass on #109 (although it isn't a *complete* solution
due to the above issue with Tonic).