docs: add Netlify auto-publishing of `main` API docs (#116)
This branch configures Netlify to automatically build and publish the
API docs for the `main` branch. This is particularly useful right now,
since the console hasn't been published to `crates.io`, so we don't have
docs.rs docs to refer people to yet. A deploy preview is also generated
for every pull request against `main`, so we can preview the new docs
added in PRs.

Since I configured the Netlify docs build to deny RustDoc warnings, I
also had to fix a couple broken links in the current docs.

The `console` crate's docs are built with private items included. This
is because the `console` application is a binary rather than a library, and
will probably never be published as a lib. Those docs should be thought
of as internal development docs, rather than user-facing API reference
docs.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>