chore(subscriber): update `tracing-subscriber` to 0.2.19
`console-subscriber` now uses the `Context::event_span` API method in
`tracing-subscriber`, which was added in version 0.2.19. However, our
minimum required version of `tracing-subscriber` in the
`console-subscriber` `Cargo.toml` is only 0.2.17. This means that cargo
is _permitted_ to select a version that may not have the required API.

This branch updates the minimum version to 0.2.19, ensuring the APIs we
use are always there.